Update transformers.py
TypeError: '>=' not supported between instances of 'str' and 'int' in line 455.modification was made on the line with the .filter(pl.col("group_ind") >= 0) operation. This fix should resolve the TypeError by correctly filtering the Polars DataFrame based on the condition where the values in the "group_ind" column are greater than or equal to 0.